Further modularize: extract sidebar resize/toggle and in-app tour into ui/sidebar.js and ui/tour.js; wire imports; update SW cache
