+++ script.js
@@ -2,6 +2,8 @@ import './pwa.js';
 import { generateId, formatDateName, isExcludedScriptName, resolveAssetPath, normalizeKey } from './utils.js';
 import { positionTooltip, showTouchAbilityPopup, positionInfoIcons } from './ui/tooltip.js';
 import { createCurvedLabelSvg, createDeathRibbonSvg } from './ui/svg.js';
+import { initSidebarResize, initSidebarToggle } from './ui/sidebar.js';
+import { initInAppTour } from './ui/tour.js';
 
 document.addEventListener('DOMContentLoaded', () => {
   const startGameBtn = document.getElementById('start-game');
@@ -2280,141 +2282,20 @@ document.addEventListener('DOMContentLoaded', () => {
     autoLoadTokens();
   }
 
-  // Sidebar resizer: drag to adjust width
-  (function initSidebarResize() {
-    if (!sidebarResizer || !sidebarEl) return;
-    // Load persisted width
-    const saved = localStorage.getItem('sidebarWidthPx');
-    if (saved) {
-      document.documentElement.style.setProperty('--sidebar-width', `${Math.max(220, Math.min(parseInt(saved,10), 600))}px`);
-    }
-    let isDragging = false;
-    let startX = 0;
-    let startWidth = 0;
-    const minW = 220;
-    const maxW = 800;
-    const onMove = (e) => {
-      if (!isDragging) return;
-      const dx = e.clientX - startX;
-      const newW = Math.max(minW, Math.min(startWidth + dx, maxW));
-      document.documentElement.style.setProperty('--sidebar-width', `${newW}px`);
-    };
-    const onUp = () => {
-      if (!isDragging) return;
-      isDragging = false;
-      document.body.classList.remove('resizing');
-      document.removeEventListener('mousemove', onMove);
-      document.removeEventListener('mouseup', onUp);
-      const val = getComputedStyle(sidebarEl).width;
-      localStorage.setItem('sidebarWidthPx', parseInt(val, 10));
-    };
-    sidebarResizer.addEventListener('mousedown', (e) => {
-      isDragging = true;
-      startX = e.clientX;
-      startWidth = sidebarEl.getBoundingClientRect().width;
-      document.addEventListener('mousemove', onMove);
-      document.addEventListener('mouseup', onUp);
-      document.body.classList.add('resizing');
-    });
+  // Sidebar resizer
+  initSidebarResize(sidebarResizer, sidebarEl);
 
-    // Touch support for resizing
-    const onTouchMove = (e) => {
-      if (!isDragging) return;
-      if (e.touches && e.touches.length) {
-        e.preventDefault();
-        const dx = e.touches[0].clientX - startX;
-        const newW = Math.max(minW, Math.min(startWidth + dx, maxW));
-        document.documentElement.style.setProperty('--sidebar-width', `${newW}px`);
-      }
-    };
-    const onTouchEnd = () => {
-      if (!isDragging) return;
-      isDragging = false;
-      document.body.classList.remove('resizing');
-      document.removeEventListener('touchmove', onTouchMove);
-      document.removeEventListener('touchend', onTouchEnd);
-      const val = getComputedStyle(sidebarEl).width;
-      localStorage.setItem('sidebarWidthPx', parseInt(val, 10));
-    };
-    sidebarResizer.addEventListener('touchstart', (e) => {
-      if (!e.touches || !e.touches.length) return;
-      isDragging = true;
-      startX = e.touches[0].clientX;
-      startWidth = sidebarEl.getBoundingClientRect().width;
-      document.addEventListener('touchmove', onTouchMove, { passive: false });
-      document.addEventListener('touchend', onTouchEnd);
-      document.body.classList.add('resizing');
-    }, { passive: true });
-  })();
-
-  // Sidebar open/close toggle with persistence
-  (function initSidebarToggle() {
-    if (!sidebarToggleBtn || !sidebarEl) return;
-    const COLLAPSE_KEY = 'sidebarCollapsed';
-    const applyCollapsed = (collapsed) => {
-      document.body.classList.toggle('sidebar-collapsed', collapsed);
-      const useOverlay = prefersOverlaySidebar.matches;
-      document.body.classList.toggle('sidebar-open', !collapsed && useOverlay);
-      if (sidebarBackdrop) sidebarBackdrop.style.display = (!collapsed && useOverlay) ? 'block' : 'none';
-      // Button label + visibility
-      sidebarToggleBtn.textContent = 'Open Sidebar';
-      sidebarToggleBtn.style.display = collapsed ? 'inline-block' : 'none';
-      sidebarToggleBtn.setAttribute('aria-pressed', String(!collapsed));
-      // Save state
-      localStorage.setItem(COLLAPSE_KEY, collapsed ? '1' : '0');
-      // Trigger layout recalculation
-      requestAnimationFrame(() => repositionPlayers());
-    };
-    // Initialize from stored state
-    const stored = localStorage.getItem(COLLAPSE_KEY);
-    const startCollapsed = stored === '1' || prefersOverlaySidebar.matches; // default collapsed on small screens
-    applyCollapsed(startCollapsed);
-    // Open button (in grimoire)
-    sidebarToggleBtn.addEventListener('click', () => {
-      applyCollapsed(false);
-    });
-    // Close button (in sidebar)
-    if (sidebarCloseBtn) {
-      sidebarCloseBtn.addEventListener('click', () => {
-        applyCollapsed(true);
-      });
-    }
-    // Backdrop click closes
-    if (sidebarBackdrop) {
-      sidebarBackdrop.addEventListener('click', () => applyCollapsed(true));
-    }
-    // React to viewport changes
-    prefersOverlaySidebar.addEventListener('change', () => {
-      const collapsed = document.body.classList.contains('sidebar-collapsed');
-      applyCollapsed(collapsed);
-    });
-    
-    // Add outside click handler for touch devices when not using overlay
-    if (isTouchDevice) {
-      const handleOutsideClick = (event) => {
-        const useOverlay = prefersOverlaySidebar.matches;
-        if (useOverlay) return; // overlay/backdrop handles closing on mobile
-        // Check if sidebar is currently open
-        if (document.body.classList.contains('sidebar-collapsed')) {
-          return; // Sidebar is already closed
-        }
-        
-        // Check if click is outside sidebar and sidebar resizer
-        const clickedInSidebar = sidebarEl.contains(event.target);
-        const clickedOnResizer = sidebarResizer && sidebarResizer.contains(event.target);
-        const clickedOnToggleButton = sidebarToggleBtn.contains(event.target);
-        
-        if (!clickedInSidebar && !clickedOnResizer && !clickedOnToggleButton) {
-          // Click is outside sidebar, close it
-          applyCollapsed(true);
-        }
-      };
-      
-      // Add both click and touchstart event listeners for better touch support
-      document.addEventListener('click', handleOutsideClick, true);
-      document.addEventListener('touchstart', handleOutsideClick, { passive: true, capture: true });
-    }
-  })();
+  // Sidebar toggle
+  initSidebarToggle({
+    sidebarToggleBtn,
+    sidebarCloseBtn,
+    sidebarBackdrop,
+    sidebarEl,
+    sidebarResizer,
+    prefersOverlaySidebar,
+    isTouchDevice,
+    repositionPlayers
+  });
 
   // Load histories and render lists
   loadHistories();
@@ -2424,319 +2305,8 @@ document.addEventListener('DOMContentLoaded', () => {
   // Restore previous session (script and grimoire)
   loadAppState();
 
-  // Simple in-app tour system
-  (function initInAppTour() {
-    const startButton = document.getElementById('start-tour');
-    const sidebar = document.getElementById('sidebar');
-    const sidebarToggleBtn = document.getElementById('sidebar-toggle');
-
-    if (!startButton || !sidebar || !sidebarToggleBtn) return;
-
-    // Create overlay elements once
-    const backdrop = document.createElement('div');
-    backdrop.className = 'tour-backdrop';
-    const highlight = document.createElement('div');
-    highlight.className = 'tour-highlight';
-    const pop = document.createElement('div');
-    pop.className = 'tour-popover';
-    pop.setAttribute('role', 'dialog');
-    pop.setAttribute('aria-live', 'polite');
-    document.body.appendChild(backdrop);
-    document.body.appendChild(highlight);
-    document.body.appendChild(pop);
-
-    // Helper to read current sidebar collapsed state
-    const isSidebarCollapsed = () => document.body.classList.contains('sidebar-collapsed');
-    // Helper to open/close sidebar programmatically using existing logic
-    function setSidebarCollapsed(collapsed) {
-      const stored = localStorage.getItem('sidebarCollapsed');
-      const currently = stored === '1';
-      // If desired state differs from effective class, toggle via UI to ensure all side effects
-      const classCollapsed = isSidebarCollapsed();
-      if (collapsed && !classCollapsed) {
-        const closeBtn = document.getElementById('sidebar-close');
-        if (closeBtn) closeBtn.click();
-        else document.body.classList.add('sidebar-collapsed');
-      } else if (!collapsed && classCollapsed) {
-        sidebarToggleBtn.click();
-      }
-    }
-
-    // Utility to ensure element is visible and sidebar state appropriate
-    function ensureVisibilityForStep(step) {
-      if (step.requiresSidebarOpen) {
-        setSidebarCollapsed(false);
-      } else if (step.requiresSidebarClosed) {
-        setSidebarCollapsed(true);
-      }
-    }
-
-    // Wait for ongoing CSS transitions/animations to complete before measuring
-    function waitForAnimationsToFinish(element, fallbackMs = 400) {
-      return new Promise((resolve) => {
-        let resolved = false;
-        function finish() {
-          if (resolved) return;
-          resolved = true;
-          resolve();
-        }
-        try {
-          const nodes = [document.body, sidebar, element].filter(Boolean);
-          const animations = [];
-          for (let i = 0; i < nodes.length; i += 1) {
-            const node = nodes[i];
-            if (node && typeof node.getAnimations === 'function') {
-              const arr = node.getAnimations({ subtree: true });
-              for (let j = 0; j < arr.length; j += 1) animations.push(arr[j]);
-            }
-          }
-          if (animations.length > 0) {
-            const timeoutId = setTimeout(finish, Math.max(250, fallbackMs));
-            Promise.all(animations.map((a) => a.finished.catch(() => {}))).then(() => {
-              clearTimeout(timeoutId);
-              // Wait two frames to allow layout to fully settle
-              requestAnimationFrame(() => requestAnimationFrame(finish));
-            });
-          } else {
-            setTimeout(finish, Math.max(250, fallbackMs));
-          }
-        } catch(_) {
-          setTimeout(finish, Math.max(250, fallbackMs));
-        }
-      });
-    }
-
-    // Compute popover position near target rect
-    function positionPopoverNear(rect) {
-      const margin = 12;
-      const popRect = pop.getBoundingClientRect();
-      let top = rect.bottom + margin;
-      let left = rect.left;
-      if (left + popRect.width > window.innerWidth - 8) {
-        left = Math.max(8, window.innerWidth - popRect.width - 8);
-      }
-      if (top + popRect.height > window.innerHeight - 8) {
-        top = Math.max(8, rect.top - popRect.height - margin);
-      }
-      pop.style.top = `${Math.max(8, top)}px`;
-      pop.style.left = `${Math.max(8, left)}px`;
-    }
-
-    // Highlight a DOM rect
-    function showHighlight(rect) {
-      const radius = 12;
-      highlight.style.display = 'block';
-      highlight.style.left = `${rect.left - 8}px`;
-      highlight.style.top = `${rect.top - 8}px`;
-      highlight.style.width = `${rect.width + 16}px`;
-      highlight.style.height = `${rect.height + 16}px`;
-      highlight.style.borderRadius = `${radius}px`;
-    }
-
-    function hideUI() {
-      backdrop.style.display = 'none';
-      highlight.style.display = 'none';
-      pop.style.display = 'none';
-    }
-
-    // Steps definition
-    const steps = [
-      {
-        id: 'welcome',
-        title: 'Quick Tour',
-        body: 'Learn the basics: set players, load a script, assign characters, add reminders. Use Next/Back or ←/→. Press Esc to exit.',
-        target: () => document.getElementById('sidebar-toggle'),
-        requiresSidebarClosed: true
-      },
-      {
-        id: 'open-sidebar',
-        title: 'Open the sidebar',
-        body: 'Open the sidebar to set up and load a script.',
-        target: () => document.getElementById('sidebar-toggle'),
-        requiresSidebarClosed: true,
-        onBeforeNext: () => setSidebarCollapsed(false)
-      },
-      {
-        id: 'game-setup',
-        title: 'Set players',
-        body: 'Choose the player count and press Start Game to create tokens.',
-        target: () => document.getElementById('start-game'),
-        requiresSidebarOpen: true
-      },
-      {
-        id: 'scripts',
-        title: 'Load a script',
-        body: 'Load a built-in script to populate roles.',
-        target: () => document.querySelector('#sidebar .script-buttons') || document.getElementById('load-status'),
-        requiresSidebarOpen: true
-      },
-      {
-        id: 'assign-character',
-        title: 'Assign a character',
-        body: 'Tap a player token to choose and assign a character.',
-        target: () => {
-          const li = document.querySelector('#player-circle li .player-token');
-          return li || document.getElementById('player-circle');
-        },
-        requiresSidebarClosed: true,
-        onEnter: () => setSidebarCollapsed(true)
-      },
-      {
-        id: 'reminders',
-        title: 'Reminders',
-        body: 'Use the + near a player to add a reminder token or text note.',
-        target: () => document.querySelector('#player-circle li .reminder-placeholder') || document.getElementById('player-circle'),
-        requiresSidebarClosed: true
-      },
-      {
-        id: 'offline',
-        title: 'Use it offline',
-        body: () => {
-          const isIOS = /iphone|ipad|ipod/i.test(navigator.userAgent);
-          const isMacSafari = /^((?!chrome|android).)*safari/i.test(navigator.userAgent) && /Mac/i.test(navigator.platform);
-          if (isIOS) {
-            return 'On iPhone/iPad: tap Share, then "Add to Home Screen" to install. The app works offline once installed.';
-          }
-          if (isMacSafari) {
-            return 'On Mac (Safari): in the Share menu, choose "Add to Dock" to install. The app will be available offline.';
-          }
-          return 'Install this app for offline use: on mobile, use "Add to Home Screen"; on desktop browsers, use "Install app" or create a shortcut.';
-        },
-        target: () => document.getElementById('sidebar-toggle'),
-        requiresSidebarClosed: true
-      },
-      {
-        id: 'finish',
-        title: "You're ready!",
-        body: 'You can restart this tour from the sidebar any time.',
-        target: () => document.getElementById('center'),
-        requiresSidebarClosed: true
-      }
-    ];
-
-    let idx = 0;
-
-    function renderStep() {
-      ensureVisibilityForStep(steps[idx]);
-
-      // If the step wants sidebar closed/open, ensure it immediately
-      if (steps[idx].onEnter) {
-        try { steps[idx].onEnter(); } catch(_) {}
-      }
-
-      const targetEl = steps[idx].target && steps[idx].target();
-
-      const doRender = () => {
-        const rect = targetEl ? targetEl.getBoundingClientRect() : { left: 16, top: 16, width: 300, height: 60, right: 316, bottom: 76 };
-
-        // Show overlay
-        backdrop.style.display = 'block';
-        showHighlight(rect);
-
-        // Build popover content
-        pop.innerHTML = '';
-
-        const title = document.createElement('div');
-        title.className = 'title';
-        title.textContent = steps[idx].title;
-
-        const body = document.createElement('div');
-        body.className = 'body';
-        const bodyText = (typeof steps[idx].body === 'function') ? steps[idx].body() : steps[idx].body;
-        body.textContent = bodyText;
-
-        const actions = document.createElement('div');
-        actions.className = 'actions';
-
-        const skipBtn = document.createElement('button');
-        skipBtn.className = 'button';
-        skipBtn.textContent = 'Skip';
-        skipBtn.onclick = endTour;
-
-        const prevBtn = document.createElement('button');
-        prevBtn.className = 'button';
-        prevBtn.textContent = 'Back';
-        prevBtn.disabled = idx === 0;
-        prevBtn.onclick = () => { idx = Math.max(0, idx - 1); renderStep(); };
-
-        const nextBtn = document.createElement('button');
-        nextBtn.className = 'button';
-        nextBtn.textContent = idx === steps.length - 1 ? 'Finish' : 'Next';
-        nextBtn.onclick = () => {
-          if (steps[idx].onBeforeNext) {
-            try { steps[idx].onBeforeNext(); } catch(_) {}
-          }
-          if (idx < steps.length - 1) { idx += 1; renderStep(); } else { endTour(); }
-        };
-
-        const progress = document.createElement('div');
-        progress.className = 'progress';
-        progress.textContent = `Step ${idx + 1} of ${steps.length}`;
-
-        actions.appendChild(skipBtn);
-        actions.appendChild(prevBtn);
-        actions.appendChild(nextBtn);
-
-        pop.appendChild(title);
-        pop.appendChild(body);
-        pop.appendChild(actions);
-        pop.appendChild(progress);
-
-        pop.style.display = 'block';
-        // Position after contents are added
-        positionPopoverNear(rect);
-      };
-
-      // If the target is inside the sidebar, scroll it into view so it's visible
-      if (targetEl && sidebar.contains(targetEl)) {
-        try { targetEl.scrollIntoView({ block: 'center', inline: 'nearest' }); } catch(_) {}
-      }
-
-      // Wait for transitions/animations to finish, then render highlight
-      waitForAnimationsToFinish(targetEl, 400).then(() => {
-        requestAnimationFrame(() => requestAnimationFrame(doRender));
-      });
-    }
-
-    function endTour() {
-      hideUI();
-      window.removeEventListener('resize', handleResize, true);
-      window.removeEventListener('orientationchange', handleResize, true);
-      document.removeEventListener('keydown', handleKey, true);
-    }
-
-    function handleResize() {
-      if (pop.style.display !== 'block') return;
-      const targetEl = steps[idx].target && steps[idx].target();
-      if (!targetEl) return;
-      const rect = targetEl.getBoundingClientRect();
-      showHighlight(rect);
-      positionPopoverNear(rect);
-    }
-
-    function handleKey(e) {
-      if (pop.style.display !== 'block') return;
-      if (e.key === 'Escape') { endTour(); }
-      if (e.key === 'ArrowRight') { const was = idx; idx = Math.min(steps.length - 1, idx + 1); if (idx !== was) renderStep(); }
-      if (e.key === 'ArrowLeft') { const was = idx; idx = Math.max(0, idx - 1); if (idx !== was) renderStep(); }
-    }
-
-    startButton.addEventListener('click', () => {
-      // Start at first step; adapt to current UI state
-      idx = 0;
-      renderStep();
-      // Add listeners for responsiveness and touch
-      window.addEventListener('resize', handleResize, true);
-      window.addEventListener('orientationchange', handleResize, true);
-      document.addEventListener('keydown', handleKey, true);
-      // Allow tapping the backdrop to advance
-      backdrop.onclick = () => {
-        if (idx < steps.length - 1) { idx += 1; renderStep(); } else { endTour(); }
-      };
-      // Prevent scrolling on touch backdrop
-      backdrop.addEventListener('touchmove', (e) => { e.preventDefault(); }, { passive: false });
-    });
-  })();
+  // In-app tour
+  initInAppTour();
 });
 
 // Tooltip and info icon helpers are imported from ui/tooltip.js
